# Conflicts:
#	src/main/java/com/quickticket/quickticket/domain/event/domain/AgeRating.java
#	src/main/java/com/quickticket/quickticket/domain/payment/credit/domain/TransactionType.java
#	src/main/java/com/quickticket/quickticket/domain/ticket/controller/TicketController.java
#	src/main/java/com/quickticket/quickticket/domain/ticket/dto/TicketResponse.java

+++ src/main/java/com/quickticket/quickticket/shared/converters/AbstractOrdinalEnumConverter.java
@@ -6,16 +6,16 @@ import lombok.AllArgsConstructor;
 import java.util.Arrays;
 
 @AllArgsConstructor
-public abstract class AbstractOrdinalEnumConverter<T extends Enum<T> & OrdinalEnum> implements AttributeConverter<T, Integer> {
+public abstract class AbstractOrdinalEnumConverter<T extends Enum<T> & OrdinalEnum> implements AttributeConverter<T, Short> {
     private final Class<T> clazz;
 
     @Override
-    public Integer convertToDatabaseColumn(T attribute) {
-        return (attribute == null) ? null : attribute.getCode();
+    public Short convertToDatabaseColumn(T attribute) {
+        return (attribute == null) ? null : ((short)attribute.getCode());
     }
 
     @Override
-    public T convertToEntityAttribute(Integer dbData) {
+    public T convertToEntityAttribute(Short dbData) {
         if (dbData == null) return null;
 
         return Arrays.stream(clazz.getEnumConstants())
